move lambdas into /value-ticker
