+++ src/value-ticker/calculate-lambda/local.ts
