+++ src/value-ticker/query-lambda/local.ts
